Unbreak linter (6/n)
### Summary

This linter fix diff focuses on the server side. I removed some obvious bad debugging statements and some testing code that should never be there.

This will probably be the last linter fix diff in this series for a while. My experience of navigating the codebase tells me that typescript should probably be introduced so that we can have less surprises about our code quality.

### Test Plan

:eyes: